Separate user settings from user-defined areas and update endpoints
